+++ docs/changelog.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## Version 0.3 [2024-04-08]
+
+- [#81](https://github.com/guibranco/pancake/issues/81) - Add Health Checks class by [@guibranco](https://github.com/guibranco).
+
 ## Version 0.2 [2024-03-20]
 
-- [#86](https://github.com/guibranco/Pancake/issues/86) - Add Request class by [@guibranco](https;//github.com/guibranco).
+- [#86](https://github.com/guibranco/pancake/issues/86) - Add Request class by [@guibranco](https;//github.com/guibranco).
 
 ## Version 0.1 [2024-03-18]